refactor: put sympy implementation in isolated sub-module
